+++ CHANGELOG.md
@@ -28,6 +28,7 @@
   - `unregescape`
   - `safeApply`
 - Removed the `warning` directive – use the `korp-warning` class directly instead.
+- Removed the unused `toBody` directive
 - Renamed localization functions (just like the template filters):
   - `getLocaleString` to `loc`
   - `getLocaleStringObject` to `locObj`